requirements.txt
pystac-client
shapely
rasterio
rioxarray
matplotlib
geopandas[all]
xarray-spatial
dask
odc-stac